Create TokenBalanceContext
