+++ src/components/Global/TokenSelect/TokenSelect.tsx
@@ -7,7 +7,7 @@ import { toDisplayQty } from '@crocswap-libs/sdk';
 import { getFormattedNumber } from '../../../App/functions/getFormattedNumber';
 import { useContext } from 'react';
 import { UserDataContext } from '../../../contexts/UserDataContext';
-import { ChainDataContext } from '../../../contexts/ChainDataContext';
+import { TokenBalanceContext } from '../../../contexts/TokenBalanceContext';
 
 interface propsIF {
     token: TokenIF;
@@ -19,7 +19,7 @@ export default function TokenSelect(props: propsIF) {
     const { token, chooseToken, fromListsText } = props;
 
     const { isUserConnected } = useContext(UserDataContext);
-    const { tokenBalances } = useContext(ChainDataContext);
+    const { tokenBalances } = useContext(TokenBalanceContext);
 
     const isMatchingToken = (tokenInRtk: TokenIF) =>
         tokenInRtk.address.toLowerCase() === token.address.toLowerCase();
