+++ src/contexts/ChainDataContext.tsx
@@ -29,14 +29,6 @@ interface ChainDataContextIF {
     lastBlockNumber: number;
     setLastBlockNumber: Dispatch<SetStateAction<number>>;
     client: Client;
-
-    tokenBalances: TokenIF[] | undefined;
-    resetTokenBalances: () => void;
-    setTokenBalance: (params: {
-        tokenAddress: string;
-        walletBalance?: string | undefined;
-        dexBalance?: string | undefined;
-    }) => void;
 }
 
 export const ChainDataContext = createContext<ChainDataContextIF>(
@@ -62,10 +54,6 @@ export const ChainDataContextProvider = (props: {
 
     const [lastBlockNumber, setLastBlockNumber] = useState<number>(0);
     const [gasPriceInGwei, setGasPriceinGwei] = useState<number | undefined>();
-
-    useEffect(() => {
-        resetTokenBalances();
-    }, [isUserConnected, userAddress]);
 
     async function pollBlockNum(): Promise<void> {
         // if default RPC is Infura, use key from env variable
@@ -170,38 +158,6 @@ export const ChainDataContextProvider = (props: {
         () => Math.floor(lastBlockNumber / 8),
         [lastBlockNumber],
     );
-    const setTokenBalance = (params: {
-        tokenAddress: string;
-        walletBalance?: string | undefined;
-        dexBalance?: string | undefined;
-    }) => {
-        if (!tokenBalances) return;
-        const newTokenBalances = [...tokenBalances];
-
-        const tokenIndex = newTokenBalances?.findIndex(
-            (token) =>
-                token.address.toLowerCase() ===
-                params.tokenAddress.toLowerCase(),
-        );
-
-        if (newTokenBalances && tokenIndex && tokenIndex !== -1) {
-            const newTokenBalance = newTokenBalances[tokenIndex];
-            if (params.walletBalance) {
-                newTokenBalance.walletBalance = params.walletBalance;
-            }
-            if (params.dexBalance) {
-                newTokenBalance.dexBalance = params.dexBalance;
-            }
-            if (params.dexBalance || params.walletBalance) {
-                newTokenBalances[tokenIndex] = newTokenBalance;
-                setTokenBalances(newTokenBalances);
-            }
-        }
-    };
-
-    const resetTokenBalances = () => {
-        setTokenBalances(undefined);
-    };
 
     useEffect(() => {
         (async () => {
@@ -253,9 +209,6 @@ export const ChainDataContextProvider = (props: {
         gasPriceInGwei,
         setGasPriceinGwei,
         client,
-        tokenBalances,
-        resetTokenBalances,
-        setTokenBalance,
     };
 
     return (
