+++ src/contexts/GlobalContexts.tsx
@@ -17,6 +17,7 @@ import { ExploreContextProvider } from './ExploreContext';
 import { ENSAddressContextProvider } from './ENSAddressContext';
 import { GraphDataContextProvider } from './GraphDataContext';
 import { DataLoadingContextProvider } from './DataLoadingContext';
+import { TokenBalanceContextProvider } from './TokenBalanceContext';
 
 export const GlobalContexts = (props: { children: React.ReactNode }) => {
     return (
@@ -25,35 +26,37 @@ export const GlobalContexts = (props: { children: React.ReactNode }) => {
                 <CachedDataContextProvider>
                     <DataLoadingContextProvider>
                         <UserDataContextProvider>
-                            <CrocEnvContextProvider>
-                                <TokenContextProvider>
-                                    <ChainDataContextProvider>
-                                        <GraphDataContextProvider>
-                                            <ChartContextProvider>
-                                                <RangeContextProvider>
-                                                    <TradeTokenContextProvider>
-                                                        <PoolContextProvider>
-                                                            <CandleContextProvider>
-                                                                <TradeTableContextProvider>
-                                                                    <UserPreferenceContextProvider>
-                                                                        <SidebarContextProvider>
-                                                                            <ExploreContextProvider>
-                                                                                {
-                                                                                    props.children
-                                                                                }
-                                                                            </ExploreContextProvider>
-                                                                        </SidebarContextProvider>
-                                                                    </UserPreferenceContextProvider>
-                                                                </TradeTableContextProvider>
-                                                            </CandleContextProvider>
-                                                        </PoolContextProvider>
-                                                    </TradeTokenContextProvider>
-                                                </RangeContextProvider>
-                                            </ChartContextProvider>
-                                        </GraphDataContextProvider>
-                                    </ChainDataContextProvider>
-                                </TokenContextProvider>
-                            </CrocEnvContextProvider>
+                            <TokenBalanceContextProvider>
+                                <CrocEnvContextProvider>
+                                    <TokenContextProvider>
+                                        <ChainDataContextProvider>
+                                            <GraphDataContextProvider>
+                                                <ChartContextProvider>
+                                                    <RangeContextProvider>
+                                                        <TradeTokenContextProvider>
+                                                            <PoolContextProvider>
+                                                                <CandleContextProvider>
+                                                                    <TradeTableContextProvider>
+                                                                        <UserPreferenceContextProvider>
+                                                                            <SidebarContextProvider>
+                                                                                <ExploreContextProvider>
+                                                                                    {
+                                                                                        props.children
+                                                                                    }
+                                                                                </ExploreContextProvider>
+                                                                            </SidebarContextProvider>
+                                                                        </UserPreferenceContextProvider>
+                                                                    </TradeTableContextProvider>
+                                                                </CandleContextProvider>
+                                                            </PoolContextProvider>
+                                                        </TradeTokenContextProvider>
+                                                    </RangeContextProvider>
+                                                </ChartContextProvider>
+                                            </GraphDataContextProvider>
+                                        </ChainDataContextProvider>
+                                    </TokenContextProvider>
+                                </CrocEnvContextProvider>
+                            </TokenBalanceContextProvider>
                         </UserDataContextProvider>
                     </DataLoadingContextProvider>
                 </CachedDataContextProvider>
