+++ src/contexts/TokenContext.tsx
@@ -7,7 +7,7 @@ import { tokenMethodsIF, useTokens } from '../App/hooks/useTokens';
 import { useTokenSearch } from '../App/hooks/useTokenSearch';
 import { TokenIF } from '../utils/interfaces/TokenIF';
 import { CrocEnvContext } from './CrocEnvContext';
-import { ChainDataContext } from './ChainDataContext';
+import { TokenBalanceContext } from './TokenBalanceContext';
 
 interface TokenContextIF {
     tokens: tokenMethodsIF;
@@ -25,8 +25,8 @@ export const TokenContext = createContext<TokenContextIF>({} as TokenContextIF);
 
 export const TokenContextProvider = (props: { children: React.ReactNode }) => {
     const { chainData } = useContext(CrocEnvContext);
-    const { tokenBalances } = useContext(ChainDataContext); // This can't come from here since it's higher order than chainContext
-
+    // TODO: possible option to merge TokenBalanceContext with TokenContext
+    const { tokenBalances } = useContext(TokenBalanceContext);
     const tokens: tokenMethodsIF = useTokens(chainData.chainId);
     const { addRecentToken, getRecentTokens } = useRecentTokens(
         chainData.chainId,
